work on manage_users.php

+++ includes/admin_tools.php
@@ -69,22 +69,23 @@ function update_all_stuff_admin($dbc){
   $results = mysqli_query($dbc , $query);
   check_results($results);
 
-  while ( $row = mysqli_fetch_array($results , MYSQLI_ASSOC))
-  {
+    if($results) {
+        while ($row = mysqli_fetch_array($results, MYSQLI_ASSOC)) {
 
-    for($i = 0; $i<mysqli_num_rows($results); $i++){
-      $status = $_POST['item-status' . $i];
-      $id = $_POST['id' . $i];
-      if($status == 'Remove'){
-       $updateQuery = 'DELETE FROM stuff WHERE id=\'' . $id . '\'';  
-      } else {
-      $updateQuery = 'UPDATE stuff SET status =\'' . $status . '\' WHERE id=\'' . $id . '\'';
-      }
-      //echo $updateQuery;
-      $results2 = mysqli_query($dbc , $updateQuery);
-      check_results($results);
+            for ($i = 0; $i < mysqli_num_rows($results); $i++) {
+                $status = $_POST['item-status' . $i];
+                $id = $_POST['id' . $i];
+                if ($status == 'Remove') {
+                    $updateQuery = 'DELETE FROM stuff WHERE id=\'' . $id . '\'';
+                } else {
+                    $updateQuery = 'UPDATE stuff SET status =\'' . $status . '\' WHERE id=\'' . $id . '\'';
+                }
+                //echo $updateQuery;
+                $results2 = mysqli_query($dbc, $updateQuery);
+                check_results($results2);
+            }
+        }
     }
-  }
 }
 
 function show_all_stuff_admin($dbc) {
@@ -165,4 +166,82 @@ function change_password($dbc, $admin_id, $newpass){
 
     //echo $query;
 }
+
+# Checks to see if an admin should be deleted.
+function update_users($dbc){
+    $query = 'SELECT * FROM users';
+
+    $results = mysqli_query($dbc, $query);
+    check_results($results);
+
+    if($results){
+        while ($row = mysqli_fetch_array($results, MYSQLI_ASSOC)) {
+            for ($i = 0; $i < mysqli_num_rows($results); $i++){
+                $buttonPressed = $_POST['delete' . $i];
+                $id = $_POST['admin_id' . $i];
+
+
+                if(isset($buttonPressed)){
+                    $deleteQuery = 'DELETE FROM users WHERE id=\'' . $id . '\'';
+                    $results_delete = mysqli_query($dbc, $deleteQuery);
+                    check_results($results_delete);
+                }
+            }
+        }
+    }
+
+
+}
+
+# Shows admin users for a particular admin to manage.
+function show_users($dbc) {
+    # Create a query to show all users
+    $query = 'SELECT user_id, username, first_name, last_name, email, reg_date FROM users ORDER BY reg_date DESC';
+
+    # Execute the query
+    $results = mysqli_query($dbc , $query);
+    check_results($results) ;
+
+    # Show results
+    if( $results )
+    {
+        # But...wait until we know the query succeed before
+        # rendering the table start.
+        echo '<p> Current Adminstrators: </p>';
+        echo '<br>';
+        echo '<TABLE class="users">';
+        echo '<TR>';
+        echo '<TH>Username</TH>';
+        echo '<TH>Name</TH>';
+        echo '<TH>E-Mail</TH>';
+        echo '<TH>Register Date</TH>';
+        echo '<TH>Options</TH>';
+        echo '</TR>';
+
+        # For each row result, generate a table row
+
+        # Increment for post variable
+        $num = 0;
+        while ( $row = mysqli_fetch_array( $results , MYSQLI_ASSOC ) )
+        {
+
+            echo '<TR>' ;
+            echo '<TD>' . $row['username'] . '</TD>';
+            echo '<TD>' . $row['first_name'] . ' ' . $row['last_name'] . '</TD>' ;
+            echo '<TD>' . $row['email'] . '</TD>' ;
+            echo '<TD>' . $row['reg_date'] . '</TD>' ;
+            echo '<TD>' . '<button onclick="return confirm(\'Are you sure you wish to delete this admin?\')" type="submit" name="delete' . $num . '">Delete User</button>' . '</TD>' ;
+            echo '</TR>';
+            echo '<input type="hidden" name="admin_id' . $num . '" value="' . $row['user_id'] . '">';
+            $num++;
+        }
+
+        # End the table
+        echo '</TABLE>';
+
+        # Free up the results in memory
+        mysqli_free_result( $results ) ;
+    }
+}
+
 ?>
